% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demenumber.R
\name{DemeNumber}
\alias{DemeNumber}
\title{This function calculate the deme number for EEMS based on the optimized edge length calculated from function Edgelength()}
\usage{
DemeNumber(outlinefile, max_edge_length)
}
\arguments{
\item{outlinefile}{outline file}

\item{max_edge_length}{optimized edge length calculated from function Edgelength()}
}
\description{
This function calculate the deme number for EEMS based on the optimized edge length calculated from function Edgelength()
}
\examples{
DemeNumber()
}
\keyword{GridGenerator}
