% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demenumber.R
\name{Edgelength}
\alias{Edgelength}
\title{This function allows you to calculate the maximum edge length for each triangle side
Edgelength()}
\usage{
Edgelength(Gene_diff, Sample_location, Automode)
}
\arguments{
\item{Gene_diff}{genetic distance file (.diffs)}

\item{Sample_location}{sample coordinates file (.coord)}

\item{Automode}{parameter to determine whether to use the defualt settings of density clustering, TRUE or FALSE}
}
\description{
This function allows you to calculate the maximum edge length for each triangle side
Edgelength()
}
\examples{
Edgelength()
}
\keyword{Edgelength}
